@@ -3,26 +3,33 @@
   "version": "1.0.0",
   "private": true,
   "main": "index.js",
+  "scripts": {
+    "build": "webpack"
+  },
   "devDependencies": {
     "babel-cli": "^6.6.5",
     "babel-core": "^6.7.2",
-    "babel-loader": "^6.2.4",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-es2015-classes": "^6.14.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.7.0",
     "babel-plugin-transform-object-assign": "^6.5.0",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "nornj": "^0.4.2-beta.27",
-    "nornj-loader": "^0.4.2-beta.1",
-    "nornj-react": "^0.4.2-beta.11",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^4.4.1",
-    "react-router": "^2.0.1",
-    "react-router-redux": "^3.0.0",
-    "redux": "^3.3.1",
-    "webpack": "^2.2.1"
+    "core-js": "^2.5.3",
+    "history": "^4.7.2",
+    "nornj": "^0.4.2-rc.34",
+    "nornj-loader": "^0.4.2-rc.1",
+    "nornj-react": "^0.4.2-rc.2",
+    "prop-types": "^15.6.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-redux": "^5.0.7",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.9",
+    "redux": "^3.7.2",
+    "webpack": "^4.1.1",
+    "webpack-cli": "^2.0.12"
   }
 }
