@@ -1,22 +1,34 @@
 {
   "name": "nornj",
-  "version": "0.4.2-rc.13",
-  "description": "A multipurpose javascript template engine",
-  "main": "lib/base.js",
+  "version": "0.4.4",
+  "description": "A multipurpose javascript template engine, support both the rendering string and the react component.",
+  "main": "dist/nornj.common.js",
+  "module": "dist/nornj.esm.js",
+  "unpkg": "dist/nornj.js",
   "files": [
     "dist/",
     "lib/",
     "src/",
     "tools/",
+    "CHANGELOG.md",
     "LICENSE",
     "README.md"
   ],
   "scripts": {
-    "umd": "cross-env BABEL_ENV=webpack webpack && npm run lib",
-    "min": "cross-env BABEL_ENV=webpack NODE_ENV=production webpack && npm run lib",
+    "cjs": "cross-env NODE_ENV=cjs rollup -c -o dist/nornj.common.js",
+    "esm": "cross-env NODE_ENV=es rollup -c -o dist/nornj.esm.js",
+    "umd": "cross-env NODE_ENV=development rollup -c -o dist/nornj.js",
+    "min": "cross-env NODE_ENV=production rollup -c -o dist/nornj.min.js",
+    "runtime-cjs": "cross-env TYPE=runtime NODE_ENV=cjs rollup -c -o dist/nornj.runtime.common.js",
+    "runtime-esm": "cross-env TYPE=runtime NODE_ENV=es rollup -c -o dist/nornj.runtime.esm.js",
+    "runtime-umd": "cross-env TYPE=runtime NODE_ENV=development rollup -c -o dist/nornj.runtime.js",
+    "runtime-min": "cross-env TYPE=runtime NODE_ENV=production rollup -c -o dist/nornj.runtime.min.js",
     "lib": "babel src --out-dir lib",
+    "build": "npm run cjs && npm run esm && npm run umd && npm run min && npm run runtime-cjs && npm run runtime-esm && npm run runtime-umd && npm run runtime-min && npm run lib",
     "lint": "eslint src",
-    "test": "karma start"
+    "test": "karma start",
+    "bdd": "jasmine JASMINE_CONFIG_PATH=test/jasmine.json",
+    "ci": "npm run test && codecov test/coverage/cobertura-coverage.xml"
   },
   "keywords": [
     "nornj",
@@ -28,6 +40,9 @@
     "type": "git",
     "url": "https://github.com/joe-sky/nornj.git"
   },
+  "bugs": {
+    "url": "https://github.com/joe-sky/nornj/issues"
+  },
   "author": "Joe_Sky <hitomi_doax@sina.com>",
   "license": "MIT",
   "devDependencies": {
@@ -35,16 +50,18 @@
     "babel-core": "^6.26.0",
     "babel-eslint": "^8.0.1",
     "babel-loader": "^7.1.2",
+    "babel-plugin-external-helpers": "^6.22.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
     "babel-register": "^6.18.0",
+    "codecov": "^3.0.0",
+    "create-react-class": "^15.6.3",
     "cross-env": "^5.1.0",
     "eslint": "^4.9.0",
-    "gulp": "^3.9.1",
-    "gulp-jasmine": "^2.1.0",
-    "gulp-notify": "^2.2.0",
     "handlebars": "^4.0.5",
     "istanbul-instrumenter-loader": "^3.0.0",
-    "jsdom": "^8.1.0",
+    "jasmine": "^2.99.0",
     "karma": "^1.5.0",
     "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "^1.1.0",
@@ -53,9 +70,15 @@
     "karma-spec-reporter": "0.0.30",
     "karma-webpack": "^2.0.5",
     "lodash": "^4.13.1",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "webpack": "^3.8.1",
-    "yargs": "^4.2.0"
+    "moment": "^2.20.1",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "rollup": "^0.55.5",
+    "rollup-plugin-babel": "^3.0.3",
+    "rollup-plugin-filesize": "^1.5.0",
+    "rollup-plugin-license": "^0.5.0",
+    "rollup-plugin-replace": "^2.0.0",
+    "rollup-plugin-uglify": "^3.0.0",
+    "webpack": "^3.11.0"
   }
 }
