@@ -1,6 +1,6 @@
 ﻿//Set default data for first render NornJ html template.
 njr.setInitialData({
   store: store,
-  hashHistory: ReactRouter.hashHistory,
+  history: _history,
   App: App
 });
