@@ -1,9 +1,5 @@
 ﻿nj.registerComponent({
-  Provider: ReactRedux.Provider,
-  Router: ReactRouter.Router,
-  Route: ReactRouter.Route,
-  Redirect: ReactRouter.Redirect,
-  'router-Link': ReactRouter.Link
+  ConnectedRouter: ReactRouterRedux.ConnectedRouter
 });
 
 nj.registerFilter({
