@@ -1,9 +1,8 @@
-﻿var reducer = Redux.combineReducers(nj.assign({}, todoApp, {
-  routing: ReactRouterRedux.routeReducer
-}));
-
-var reduxRouterMiddleware = ReactRouterRedux.syncHistory(ReactRouter.hashHistory);
-var createStoreWithMiddleware = Redux.applyMiddleware(reduxRouterMiddleware)(Redux.createStore);
-
-var store = createStoreWithMiddleware(reducer);
-reduxRouterMiddleware.listenForReplays(store);
+﻿var _history = History.createHashHistory();
+var middleware = ReactRouterRedux.routerMiddleware(_history);
+var store = Redux.createStore(
+  Redux.combineReducers(nj.assign({}, todoApp, {
+    routing: ReactRouterRedux.routerReducer
+  })),
+  Redux.applyMiddleware(middleware)
+);
