@@ -1,4 +1,5 @@
-﻿import { Component, PropTypes } from 'react';
+﻿import { Component } from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { registerComponent } from 'nornj';
 import tmpls from '../template.nj.html';
