@@ -9,6 +9,13 @@ export const filters = {
     if (obj == null) {
       return obj;
     }
+    if (obj._njCtx) {
+      return {
+        _njCtx: obj.val,
+        val: obj.val[prop],
+        prop
+      };
+    }
 
     return obj[prop];
   },
@@ -31,8 +38,16 @@ export const filters = {
 
     return getComputedData({
       val: obj[prop],
-      ctx: obj
+      _njCtx: obj
     }, options.context, options.level);
+  },
+
+  '=': (obj, val) => {
+    if (obj == null) {
+      return obj;
+    }
+
+    obj._njCtx[obj.prop] = val;
   },
 
   '==': (val1, val2) => val1 == val2,
@@ -65,10 +80,10 @@ export const filters = {
 
   '**': (val1, val2) => Math.pow(val1, val2),
 
-  '//': val => Math.sqrt(val),
+  '%%': (val1, val2) => Math.floor(val1 / val2),
 
   //Ternary operator
-  '?': (val, val1, val2) => val ? val1 : val2,
+  '?:': (val, val1, val2) => val ? val1 : val2,
 
   '!': val => !val,
 
@@ -134,7 +149,9 @@ export const filters = {
     } else {
       return -1;
     }
-  }
+  },
+
+  bracket: val => val
 };
 
 function _getArrayByNum(isContainEnd) {
@@ -146,7 +163,6 @@ function _getArrayByNum(isContainEnd) {
 function _config(params) {
   let ret = {
     onlyGlobal: false,
-    transOperator: false,
     hasOptions: true
   };
 
@@ -156,49 +172,53 @@ function _config(params) {
   return ret;
 }
 
-const _defaultCfg = { onlyGlobal: true, hasOptions: false },
-  _defaultCfgO = { onlyGlobal: true, transOperator: true, hasOptions: false };
+const _defaultCfg = { onlyGlobal: true, hasOptions: false };
 
 //Filter default config
 export const filterConfig = {
   '.': _config(_defaultCfg),
   '_': _config({ onlyGlobal: true }),
   '#': _config({ onlyGlobal: true }),
-  '==': _config(_defaultCfgO),
-  '===': _config(_defaultCfgO),
-  '!=': _config(_defaultCfgO),
-  '!==': _config(_defaultCfgO),
-  '<': _config(_defaultCfgO),
-  '<=': _config(_defaultCfgO),
-  '>': _config(_defaultCfgO),
-  '>=': _config(_defaultCfgO),
-  '+': _config(_defaultCfgO),
-  '-': _config(_defaultCfgO),
-  '*': _config(_defaultCfgO),
-  '/': _config(_defaultCfgO),
-  '%': _config(_defaultCfgO),
-  '**': _config(_defaultCfgO),
-  '//': _config(_defaultCfg),
-  '?': _config(_defaultCfgO),
+  '==': _config(_defaultCfg),
+  '===': _config(_defaultCfg),
+  '!=': _config(_defaultCfg),
+  '!==': _config(_defaultCfg),
+  '<': _config(_defaultCfg),
+  '<=': _config(_defaultCfg),
+  '>': _config(_defaultCfg),
+  '>=': _config(_defaultCfg),
+  '+': _config(_defaultCfg),
+  '-': _config(_defaultCfg),
+  '*': _config(_defaultCfg),
+  '/': _config(_defaultCfg),
+  '%': _config(_defaultCfg),
+  '**': _config(_defaultCfg),
+  '%%': _config(_defaultCfg),
+  '?:': _config(_defaultCfg),
   '!': _config(_defaultCfg),
-  '&&': _config(_defaultCfgO),
-  or: _config(_defaultCfgO),
+  '&&': _config(_defaultCfg),
+  or: _config(_defaultCfg),
   int: _config(_defaultCfg),
   float: _config(_defaultCfg),
   bool: _config(_defaultCfg),
   obj: _config(_defaultCfg),
-  ':': _config(_defaultCfgO),
+  ':': _config(_defaultCfg),
   list: _config(_defaultCfg),
   reg: _config(_defaultCfg),
   css: _config(_defaultCfg),
-  '..': _config(_defaultCfgO),
-  rLt: _config(_defaultCfgO),
-  '<=>': _config(_defaultCfgO)
+  '..': _config(_defaultCfg),
+  rLt: _config(_defaultCfg),
+  '<=>': _config(_defaultCfg),
+  bracket: _config(_defaultCfg)
 };
 
 //Filter alias
 filters.prop = filters['.'];
 filterConfig.prop = filterConfig['.'];
+filters['?'] = filters['?:'];
+filterConfig['?'] = filterConfig['?:'];
+filters['//'] = filters['%%'];
+filterConfig['//'] = filterConfig['%%'];
 
 //Register filter and also can batch add
 export function registerFilter(name, filter, options) {
@@ -221,30 +241,12 @@ export function registerFilter(name, filter, options) {
         filters[name] = v;
       }
       filterConfig[name] = _config(options);
-
-      if (options && options.transOperator) {
-        _REGEX_TRANSOPTS += '|' + name;
-        nj.regexTransOpts = _getRegexTransopts(_REGEX_TRANSOPTS);
-      }
     }
   }, false, false);
 }
 
-function _getRegexTransopts(opts) {
-  return new RegExp('[\\s]+(' + opts.replace(/\+|\*|\?|\./g, match => '\\' + match) + ')[\\s]+' + nj.regexJsBase + '([^\\s,()]*)', 'g');
-}
-
-let _REGEX_TRANSOPTS = '';
-tools.each(filterConfig, (v, k) => {
-  if (v.transOperator) {
-    _REGEX_TRANSOPTS += '|' + k;
-  }
-});
-_REGEX_TRANSOPTS = _REGEX_TRANSOPTS.substr(1);
-
 tools.assign(nj, {
   filters,
   filterConfig,
-  registerFilter,
-  regexTransOpts: _getRegexTransopts(_REGEX_TRANSOPTS)
+  registerFilter
 });
