@@ -21,6 +21,8 @@ function _plainTextNode(obj, parent, parentContent, noSplitNewline, tmplRule) {
   parent[parentContent].push(node);
 }
 
+const REGEX_REPLACE_BP = /_njBp(\d+)_/g;
+
 //检测元素节点
 export default function checkElem(obj, parent, tmplRule, hasExProps, noSplitNewline, isLast) {
   const parentContent = 'content';
@@ -28,10 +30,15 @@ export default function checkElem(obj, parent, tmplRule, hasExProps, noSplitNewl
   if (!tools.isArray(obj)) { //判断是否为文本节点
     if (tools.isString(obj)) {
       if (!noSplitNewline) {
-        let strs = obj.split(tmplRule.newlineSplit);
+        const braceParams = [];
+        let strs = obj.replace(tmplRule.braceParamG, match => {
+          braceParams.push(match);
+          return '_njBp' + (braceParams.length - 1) + '_';
+        }).split(/\n/g);
+
         strs.forEach((str, i) => {
           str = str.trim();
-          str !== '' && _plainTextNode(str, parent, parentContent, noSplitNewline, tmplRule);
+          str !== '' && _plainTextNode(str.replace(REGEX_REPLACE_BP, (all, g1) => braceParams[g1]), parent, parentContent, noSplitNewline, tmplRule);
         });
       } else {
         _plainTextNode(isLast && parent.allowNewline === 'nlElem' ? tools.trimRight(obj) : obj, parent, parentContent, noSplitNewline, tmplRule);
@@ -107,7 +114,7 @@ export default function checkElem(obj, parent, tmplRule, hasExProps, noSplitNewl
             return;
           }
 
-          let paramV = tranParam.compiledParam(value, tmplRule);
+          let paramV = tranParam.compiledParam(value, tmplRule, param.hasColon, param.onlyKey);
           if (param.onlyBrace) { //提取匿名参数
             node.args.push(paramV);
           } else {
@@ -145,7 +152,7 @@ export default function checkElem(obj, parent, tmplRule, hasExProps, noSplitNewl
           }
 
           tools.each(tagParams, param => { //The parameter like "{prop}" needs to be replaced.
-            node.params[param.onlyBrace ? param.onlyBrace.replace(/\.\.\//g, '') : param.key] = tranParam.compiledParam(param.value, tmplRule);
+            node.params[param.onlyBrace ? param.onlyBrace.replace(/\.\.\//g, '') : param.key] = tranParam.compiledParam(param.value, tmplRule, param.hasColon, param.onlyKey);
           }, false, true);
         }
 
