@@ -50,8 +50,9 @@ export function getData(prop, data, hasCtx) {
       if (ret !== undefined) {
         if (hasCtx) {
           return {
-            ctx: obj,
-            val: ret
+            _njCtx: obj,
+            val: ret,
+            prop
           };
         }
 
@@ -82,11 +83,11 @@ export function getComputedData(fn, p2, level) {
       _njIcp: p2.icp
     });
   } else { //普通函数
-    return fn.val.call(fn.ctx, p2);
+    return fn.val.call(fn._njCtx, p2);
   }
 }
 
-export function getElement(name, p1, nameO, p2) {
+export function getElement(name, p1, nameO, p2, subName) {
   let element;
   if (!p2.icp) {
     element = p1.cp[name];
@@ -95,6 +96,10 @@ export function getElement(name, p1, nameO, p2) {
     if (!element) {
       element = p1.cp[name];
     }
+  }
+
+  if (subName != null && element) {
+    element = element[subName];
   }
 
   return element ? element : name;
@@ -123,18 +128,12 @@ export function newContext(p2, p3) {
     return p2;
   }
 
-  const newData = [];
-  if ('data' in p3) {
-    newData.push(p3.data);
-  }
-  if ('extra' in p3) {
-    newData.push(p3.extra);
-  }
-
   return {
-    data: newData.length ? tools.arrayPush(newData, p2.data) : p2.data,
+    data: p3.data ? tools.arrayPush(p3.data, p2.data) : p2.data,
     parent: p3.fallback ? p2 : p2.parent,
+    root: p2.root || p2,
     index: 'index' in p3 ? p3.index : p2.index,
+    item: 'item' in p3 ? p3.item : p2.item,
     level: p2.level,
     getData,
     d: getData,
@@ -170,7 +169,7 @@ export function assignStrProps(paramsE, keys) {
 
 //创建扩展标签子节点函数
 export function exRet(p1, p2, fn, p4, p5) {
-  return function(param) {
+  return function (param) {
     return fn(p1, p2, param, p4, p5);
   };
 }
@@ -191,7 +190,7 @@ function _getLocalComponents(localConfigs, initCtx) {
 
 //构建可运行的模板函数
 export function tmplWrap(configs, main) {
-  return function(lc, lc2) {
+  return function (lc, lc2) {
     const initCtx = this,
       data = tools.arraySlice(arguments);
 
@@ -199,6 +198,7 @@ export function tmplWrap(configs, main) {
       data: initCtx && initCtx._njData ? tools.arrayPush(data, initCtx._njData) : data,
       parent: initCtx ? initCtx._njParent : null,
       index: initCtx ? initCtx._njIndex : null,
+      item: initCtx ? initCtx._njItem : null,
       level: initCtx ? initCtx._njLevel : null,
       getData,
       d: getData,
